requirements/common.txt
Django
djangorestframework
mysqlclient
django-bootstrap-v5